+++ aerich.ini
@@ -1,5 +1,4 @@
 [aerich]
-tortoise_orm = service.db.db_connect.TORTOISE_ORM
+tortoise_orm = db.db_connect.TORTOISE_ORM
 location = ./data/migrations
 src_folder = ./.
-
