+++ bot.py
@@ -4,7 +4,7 @@ from nonebot.adapters.cqhttp import Bot as CQHTTPBot
 import utils.patcher  # type: ignore
 from utils.init import init_bot_startup, update_plugin_list
 from utils.browser import close_browser
-from service.db.db_connect import db_disconnect
+from db.db_connect import db_disconnect
 
 nonebot.init()
