+++ db/models/statistic.py
