+++ plugins/bili_sub/live_pusher.py
@@ -7,7 +7,7 @@ from nonebot.plugin import require
 from nonebot_plugin_apscheduler import scheduler
 from nonebot.adapters.cqhttp.message import Message, MessageSegment
 
-from service.db.models.bs import BiliSub as DB
+from db.models.bs import BiliSub as DB
 
 from .bilireq import BiliReq
