+++ plugins/illust/__init__.py
@@ -9,10 +9,9 @@ from zhconv import convert
 from nonebot.plugin import on_shell_command
 from nonebot.typing import T_State
 from nonebot.adapters.cqhttp import Bot, MessageEvent, GroupMessageEvent
-from nonebot.adapters.cqhttp.event import GroupMessageEvent
 
 from utils.msg_util import text, image, reply
-from service.db.utils.illust_config import get_illust_config
+from db.utils.illust_config import IllustConfig
 
 from .parser import pic_parser, set_parser
 
@@ -29,7 +28,7 @@ async def _(bot: Bot, event: MessageEvent, state: T_State):
         result = await args.handle(args)
         if result:
             r = result
-            c = await get_illust_config()
+            c = await IllustConfig.get_illust_config()
             tags = convert(r["tags"], "zh-tw")
             message = (
                 f"{r['title']}({r['pid']})\n作者: {r['author']}({r['uid']})\ntags: {tags}"
@@ -99,7 +98,7 @@ async def _s(bot: Bot, event: MessageEvent, state: T_State):
     if hasattr(args, "handle"):
         result = await args.handle(args)
         if result:
-            current_config = await get_illust_config()
+            current_config = await IllustConfig.get_illust_config()
             c: list[str] = []
             for n in current_config.keys():
                 c.append(f"{n}, {'开启' if current_config[n] else '关闭'}")
