+++ plugins/stats_log/stat_hook.py
@@ -1,14 +1,12 @@
 import time
 
-from nonebot import logger
-from configs.config import HIDDEN_PLUGINS
-from nonebot.plugin import get_plugin, get_loaded_plugins
 from nonebot.typing import T_State
 from nonebot.matcher import Matcher
 from nonebot.message import run_postprocessor
-from nonebot.exception import IgnoredException
 from nonebot.adapters.cqhttp import Bot, Event, GroupMessageEvent
-from service.db.models.statistic import Statistic
+
+from configs.config import HIDDEN_PLUGINS
+from db.models.statistic import Statistic
 
 stat: dict[int, dict[int, dict[str, int]]]
 # {gid: {day1: {plugin: times}, day2: {plugin: times}}}
